Option for a source-name different from the actual source file
While absolute paths are useful for reading source files reliably, we don't
want to save those absolute paths in precompiled files as the source files
may move after precompilation. This happens when modules are packaged for
Linux distributions and the build process runs with an unprivileged user
while the result should be installed in a system location.

The source-name option allows for specifying a different string to use for
the $?FILES variable and thus for the path to the source file in backtraces.